+++ packages/biscuit-store/CHANGELOG.md
@@ -3,17 +3,17 @@
 ## 1.2.0
 ### Major Changes
 
-- Added combineActions function
-- Added addToContainer field to createStore
-- Added checking the existence of a repository by name
-- Added combineAction field to createStore
-- Refactor creator
-- Added test for new functions
+- added combineActions function
+- added addToContainer field to createStore
+- added checking the existence of a repository by name
+- added combineAction field to createStore
+- refactor creator
+- added test for new functions
 
 ## 1.1.0
 ### Major Changes
-- Added container function
-- Added initialCall function
+- added container function
+- added initialCall  function
 
 ## 1.0.0
 ### Major Changes
@@ -27,6 +27,75 @@
 - 7cca360: Remove manager.remove()
 - ec22884: remove combineStateCollection
 - 79d022f: fix docs
+
+## 0.9.99
+### Patch Changes
+
+- Added adapter unit test
+- Added call method to adapter
+- Added channel method to adapter
+- Added createActionTo unit test
+- refactor store adn actions
+
+# @biscuit-store/adapter
+
+## 1.1.0
+### Major Changes
+
+- State and payload include to context
+- Added includeContext function
+
+## 0.9.99
+### Patch Changes
+
+- Added adapter unit test
+- Added call method to adapter
+- Added channel method to adapter
+- Added createActionTo unit test
+- refactor store adn actions
+
+# @biscuit-store/react
+
+## 2.0.0
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @biscuit-store/core@2.0.0
+
+## 1.1.0
+### Major Changes
+
+- Added listen function
+- Refactor objerver function
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @biscuit-store/core@2.0.0
+
+## 1.0.0
+### Major Changes
+
+- 038a474: fix useDebounce hook
+- 038a474: fix unsubscribe to susscribe decorator
+- 038a474: add subscribe test
+- 038a474: add useDispatchDebounce hook
+- 038a474: add observer test
+- 038a474: add useSubscribe test
+- 038a474: add useDispatchThrottle test
+- 038a474: add useDispatch unit test
+
+### Patch Changes
+  - @biscuit-store/core@1.0.0
+
+## 1.0.0
+### Patch Changes
+  - @biscuit-store/core@1.0.0
 
 ## 0.9.99
 ### Patch Changes