+++ packages/types/CHANGELOG.md
@@ -1,4 +1,8 @@
-# @biscuit-store/adapter
+# @biscuit-store/types
+
+## 1.2.1
+### Patch Changes
+- Added type field addTocontainer
 
 ## 1.2.0
 ### Patch Changes